Feat: Reordered classes to have a proper handler different from the patcher class.
